feat: script that tests parser against all ranges in plan
The goal of the parser should be that every range that's in a plan could be
parsed to match the range that we've manually fixed up for our first schedule.
But, there are some parsing bugs now that prevent that from happening. As we fix
these bugs, we can add tests for them, and once they're all fixed, we can add a
test that tests every range in all our schedules every time we run the parser
tests.

For now, these are the known parsing bugs that this script exposes:

 * Verses that are followed by letter like 1a, 1b, 1c
 * Ranges that span multiple chapters and have verse (e.g. 1:4-2:1)
 * Adjacent verses or chapters separated by commas
 * Whole book names

Those four categories of things don't parse into the expected range. Run this
script for more details.